+++ ack-angular/pipes.class.d.ts
@@ -5,6 +5,8 @@ export declare function markdownAnchor(input: string): string;
 export declare function textDownload(input: any): any;
 export declare function yesno(input: any): any;
 export declare function yesNo(input: any): any;
+export declare function boolean(input: any): boolean;
+export declare function bit(input: any): 1 | 0;
 export declare function numbers(input: any): any;
 export declare function capitalizeWords(input: any): any;
 /** each sentence word is capitalized */